Move more SameDiffTests to OpValidation suite
