+++ nd4j/nd4j-backends/nd4j-tests/src/test/java/org/nd4j/autodiff/opvalidation/ShapeOpValidation.java
@@ -12,6 +12,8 @@ import org.nd4j.linalg.api.ndarray.INDArray;
 import org.nd4j.linalg.checkutil.NDArrayCreationUtil;
 import org.nd4j.linalg.factory.Nd4j;
 import org.nd4j.linalg.factory.Nd4jBackend;
+import org.nd4j.linalg.indexing.NDArrayIndex;
+import org.nd4j.linalg.ops.transforms.Transforms;
 import org.nd4j.linalg.primitives.Pair;
 import org.nd4j.linalg.primitives.Triple;
 import org.nd4j.linalg.util.ArrayUtil;
@@ -20,7 +22,9 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
+import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNull;
 import static org.nd4j.linalg.indexing.NDArrayIndex.interval;
 
 public class ShapeOpValidation extends BaseOpValidation {
@@ -636,5 +640,239 @@ public class ShapeOpValidation extends BaseOpValidation {
         }
 
         assertEquals(failed.toString(), 0, failed.size());
+    }
+
+
+    @Test
+    public void testReshape() {
+        SameDiff sameDiff = SameDiff.create();
+        INDArray arr = Transforms.sigmoid(Nd4j.linspace(-5, 6, 12)).reshape(3, 4);
+        SDVariable x = sameDiff.var("x", arr);
+        SDVariable result1 = sameDiff.reshape(x, 4, 3);
+        SDVariable loss = sameDiff.standardDeviation(result1, true);
+
+        INDArray exp = arr.dup('c').reshape('c', 4,3);
+
+        String err = OpValidation.validate(new TestCase(sameDiff)
+                .expectedOutput(result1.getVarName(), exp));
+
+        assertNull(err);
+    }
+
+    @Test
+    public void testTranspose() {
+        SameDiff sameDiff = SameDiff.create();
+        INDArray arr = Transforms.sigmoid(Nd4j.linspace(1, 4, 4));
+        SDVariable x = sameDiff.var("x", arr);
+        SDVariable result = sameDiff.transpose(x);
+        SDVariable loss = sameDiff.standardDeviation(result, true);
+
+        String err = OpValidation.validate(new TestCase(sameDiff).expectedOutput(result.getVarName(), arr.transpose()));
+        assertNull(err);
+    }
+
+    @Test
+
+    public void testShape() {
+        SameDiff sameDiff = SameDiff.create();
+        val shape = new long[]{2, 3};
+        SDVariable x = sameDiff.var("x", shape);
+        SDVariable result = sameDiff.shape(x);
+        SDVariable loss = sameDiff.standardDeviation(result, true);
+
+        String err = OpValidation.validate(new TestCase(sameDiff)
+                .expected(result, Nd4j.create(new double[]{2,3}, new long[]{2})));
+
+        assertNull(err);
+    }
+
+
+    @Test
+    public void testReverseSequence() {
+        SameDiff sameDiff = SameDiff.create();
+        float[] input_data = new float[]{
+                1, 2, 3,
+                4, 5, 6,
+                7, 8, 9,
+                0, 0, 0,
+                0, 0, 0,
+
+                1, 2, 3,
+                4, 5, 6,
+                0, 0, 0,
+                0, 0, 0,
+                0, 0, 0
+        };
+        float[] expected_output = new float[]{
+                7, 8, 9,
+                4, 5, 6,
+                1, 2, 3,
+                0, 0, 0,
+                0, 0, 0,
+
+                4, 5, 6,
+                1, 2, 3,
+                0, 0, 0,
+                0, 0, 0,
+                0, 0, 0
+        };
+        INDArray arr1 = Nd4j.create(input_data, new long[]{2, 5, 3});
+        INDArray arr2 = Nd4j.create(new float[]{3, 2}).reshape(2);
+        SDVariable x = sameDiff.var("x", arr1);
+        SDVariable seq_lengths = sameDiff.var("seq_lengths", arr2);
+        SDVariable result = sameDiff.reverseSequence(x, seq_lengths, 1, 0);
+        INDArray expected = Nd4j.create(expected_output, new long[]{2, 5, 3});
+        assertArrayEquals(arr1.shape(), result.eval().shape());
+        assertEquals(expected, result.eval());
+
+        SDVariable loss = sameDiff.standardDeviation(result, true);
+        String err = OpValidation.validate(new TestCase(sameDiff)
+                .expected(result.getVarName(), expected)
+                .gradCheckSkipVariables(seq_lengths.getVarName()));
+        assertNull(err);
+    }
+
+    @Test
+    public void testSequenceMask() {
+        SameDiff sameDiff = SameDiff.create();
+        INDArray arr = Nd4j.create(new float[] {1, 3, 2}).reshape(3);
+        SDVariable lengths = sameDiff.var("lengths", arr);
+
+        // Test with static max len
+        int maxlen = 5;
+        INDArray expected = Nd4j.create(new float[] {1, 0, 0, 0, 0,
+                        1, 1, 1, 0, 0,
+                        1, 1, 0, 0, 0},
+                new long[]{3, 5});
+        SDVariable result1 = sameDiff.sequenceMask(lengths, maxlen);
+        assertArrayEquals(expected.shape(), result1.eval().shape());
+        assertEquals(expected, result1.eval());
+
+        SDVariable loss = sameDiff.standardDeviation(result1, true);
+
+        String err = OpValidation.validate(new TestCase(sameDiff)
+                .expected(result1, expected)
+                .gradCheckSkipVariables(lengths.getVarName()));
+        assertNull(err);
+
+        // Test with dynamic maxlen
+        lengths = sameDiff.var("lengths2", arr); // required because of an internal samediff bug
+        SDVariable maxLen = sameDiff.var("maxLen", Nd4j.create(new float[]{5}).reshape(1));
+        SDVariable result2 = sameDiff.sequenceMask(lengths, maxLen);
+        assertArrayEquals(expected.shape(), result2.eval().shape());
+        assertEquals(expected, result2.eval());
+    }
+
+    //TODO UPDATE TO OPVALIDATION
+    @Test
+    public void testGather() {
+        SameDiff sameDiff = SameDiff.create();
+        INDArray arr = Nd4j.create(new float[]{1, 2, 3, 4}, new long[]{2, 2});
+        SDVariable x = sameDiff.var("x", arr);
+        SDVariable result = sameDiff.gather(x, new int[]{1, 0}, 1);
+        INDArray expected = Nd4j.create(new float[]{2, 1, 4, 3}, new long[]{2, 2});
+        assertEquals(expected, result.eval());
+    }
+
+
+
+    //TODO UPDATE TO OPVALIDATION
+    @Test
+    public void testGatherNd() {
+        SameDiff sameDiff = SameDiff.create();
+        INDArray arr1 = Transforms.sigmoid(Nd4j.linspace(1, 24, 24)).reshape(2, 3, 4);
+        INDArray arr2 = Nd4j.create(new float[]{1, 2, 3, 0, 1, 3, 1, 0, 2}, new long[]{3, 3});
+        SDVariable x = sameDiff.var("x", arr1);
+        SDVariable idxs = sameDiff.var("idxs", arr2);
+        SDVariable result = sameDiff.gatherNd(x, idxs);
+        // build expected output array
+        INDArray expected  = Nd4j.zeros(3);
+        for (int i=0; i<3; i++){
+            INDArray idx = arr2.get(NDArrayIndex.point(i));
+            expected.get(NDArrayIndex.point(i)).assign(
+                    arr1.get(NDArrayIndex.point(idx.getInt(0)),
+                            NDArrayIndex.point(idx.getInt(1)),
+                            NDArrayIndex.point(idx.getInt(2))));
+        }
+        assertEquals(expected, result.eval());
+    }
+
+    //TODO UPDATE TO OPVALIDATION
+    @Test
+    public void testStack2() {
+        SameDiff sameDiff = SameDiff.create();
+        INDArray arr1 = Transforms.sigmoid(Nd4j.linspace(1, 6, 6)).reshape(3, 2);
+        INDArray arr2 = Transforms.sigmoid(Nd4j.linspace(7, 12, 6)).reshape(3, 2);
+        SDVariable x1 = sameDiff.var("x1", arr1);
+        SDVariable x2 = sameDiff.var("x2", arr2);
+        SDVariable result = sameDiff.stack(1, x1, x2);
+        assertArrayEquals(new long[]{3, 2, 2}, result.eval().shape());
+    }
+
+    //TODO UPDATE TO OPVALIDATION
+    @Test
+    public void testParallelStack() {
+        SameDiff sameDiff = SameDiff.create();
+        INDArray arr1 = Transforms.sigmoid(Nd4j.linspace(1, 6, 6)).reshape(3, 2);
+        INDArray arr2 = Transforms.sigmoid(Nd4j.linspace(7, 12, 6)).reshape(3, 2);
+        SDVariable x1 = sameDiff.var("x1", arr1);
+        SDVariable x2 = sameDiff.var("x2", arr2);
+        SDVariable result = sameDiff.parallel_stack(new SDVariable[]{x1, x2});
+        assertArrayEquals(new long[]{2, 3, 2}, result.eval().shape());
+        assertEquals(Nd4j.concat(0, arr1, arr2).reshape(2, 3, 2), result.eval());
+    }
+
+    //TODO UPDATE TO OPVALIDATION
+    @Test
+    public void testUnStack2() {
+        Nd4j.getExecutioner().enableDebugMode(true);
+        Nd4j.getExecutioner().enableVerboseMode(true);
+        SameDiff sameDiff = SameDiff.create();
+        INDArray arr1 = Nd4j.zeros(3, 2);
+        INDArray arr2 = Nd4j.ones(3, 2);
+        SDVariable x1 = sameDiff.var("x1", arr1);
+        SDVariable x2 = sameDiff.var("x2", arr2);
+        SDVariable stacked = sameDiff.stack(0, x1, x2);
+        SDVariable[] result = sameDiff.unstack(stacked, 0, 2);
+        assertEquals(arr1, result[0].eval());
+        assertEquals(arr2, result[1].eval());
+    }
+
+    //TODO UPDATE TO OPVALIDATION
+    @Test
+    public void testPermute() {
+        SameDiff sameDiff = SameDiff.create();
+        INDArray arr = Transforms.sigmoid(Nd4j.linspace(1, 6, 6).reshape(2, 3));
+        SDVariable x = sameDiff.var("x", arr);
+        SDVariable result = sameDiff.permute(x, 1, 0);
+        assertArrayEquals(new long[]{3, 2}, result.getShape());
+
+    }
+
+    //TODO UPDATE TO OPVALIDATION
+    @Test
+    public void testConcat2() {
+        SameDiff sameDiff = SameDiff.create();
+        INDArray arr1 = Transforms.sigmoid(Nd4j.linspace(1, 4, 4));
+        INDArray arr2 = Transforms.sigmoid(Nd4j.linspace(4, 8, 4));
+        SDVariable x1 = sameDiff.var("x1", arr1);
+        SDVariable x2 = sameDiff.var("x2", arr2);
+        SDVariable result = sameDiff.concat(0, x1, x2);
+        assertArrayEquals(new long[]{2, 4}, result.eval().shape());
+
+    }
+
+    //TODO UPDATE TO OPVALIDATION
+    @Test
+    public void testTile2() {
+        SameDiff sameDiff = SameDiff.create();
+        INDArray arr = Transforms.sigmoid(Nd4j.linspace(1, 4, 4));
+        SDVariable x = sameDiff.var("x", arr);
+        SDVariable result = sameDiff.tile(x, new int[]{2, 2});
+        assertArrayEquals(new long[]{2, 8}, result.eval().shape());
+        INDArray arr2 = Nd4j.concat(0, arr, arr);  // (1, 4), (1, 4) -> (2, 4)
+        INDArray expected = Nd4j.concat(1, arr2, arr2);  // (2, 4), (2, 4) -> (2, 8)
+        assertEquals(expected, result.eval());
+
     }
 }
